# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
    "/tree": {
      "get": {
        "operationId": "get_tree",
        "parameters": [
          {
            "name": "include_debug",
            "in": "query",
            "description": "Whether to include debug state in the supervision tree",
            "required": false,
            "schema": {
              "type": [
                "boolean",
                "null"
              ]
            }
          },
          {
            "name": "pid",
            "in": "query",
            "description": "The PID from which to start the supervision tree. If not provided, the root supervisor will be used.",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/String"
                }
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupervisionTree"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ActorStatus": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "initializing"
            ]
          },
          {
            "type": "string",
            "enum": [
              "running"
            ]
          },
          {
            "type": "string",
            "enum": [
              "suspended"
            ]
          },
          {
            "type": "string",
            "enum": [
              "exiting"
            ]
          },
          {
            "type": "object",
            "required": [
              "dead"
            ],
            "properties": {
              "dead": {
                "$ref": "#/components/schemas/Exit"
              }
            }
          }
        ]
      },
      "DebugState": {
        "type": "object",
        "required": [
          "status",
          "uptime",
          "description"
        ],
        "properties": {
          "description": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ActorStatus"
          },
          "uptime": {
            "$ref": "#/components/schemas/DurationSchema"
          }
        }
      },
      "DurationSchema": {
        "type": "object",
        "required": [
          "secs",
          "nanos"
        ],
        "properties": {
          "nanos": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "secs": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        },
        "example": "{\"secs\": 5, \"nanos\": 0}"
      },
      "Exit": {
        "oneOf": [
          {
            "type": "string",
            "enum": [
              "Normal"
            ]
          },
          {
            "type": "object",
            "required": [
              "Error"
            ],
            "properties": {
              "Error": {
                "$ref": "#/components/schemas/ExitError"
              }
            }
          }
        ]
      },
      "ExitError": {
        "type": "string",
        "enum": [
          "Panic",
          "Abort",
          "UnhandledError"
        ]
      },
      "RestartMode": {
        "type": "string",
        "enum": [
          "Always",
          "OnError",
          "Never"
        ]
      },
      "String": {
        "type": "string"
      },
      "SupervisionTree": {
        "type": "object",
        "required": [
          "pid",
          "restart_mode",
          "abort_timeout"
        ],
        "properties": {
          "abort_timeout": {
            "$ref": "#/components/schemas/DurationSchema"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SupervisionTree"
            }
          },
          "debug_state": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/DebugState"
              }
            ]
          },
          "pid": {
            "$ref": "#/components/schemas/String"
          },
          "restart_mode": {
            "$ref": "#/components/schemas/RestartMode"
          }
        }
      }
    }
  }
}